Conflicts:
	GameEngine/commands.txt
	GameEngine/src/com/javagameengine/TestGame.java
	GameEngine/src/com/javagameengine/gui/GUIcomponent.java
	GameEngine/src/com/javagameengine/renderer/Renderer.java

+++ GameEngine/src/com/javagameengine/gui/Button.java
@@ -70,7 +70,7 @@ public abstract class Button extends GUIcomponent implements Listener {
 		}
 	}
 
-	public abstract void onUpdate(int delta);
+	public abstract void onUpdate(float delta);
 	public abstract void onDestroy();
 	public abstract void onCreate();
