+++ GameEngine/src/com/javagameengine/gui/GUIcomponent.java
@@ -175,7 +175,7 @@ public abstract class GUIcomponent {
 		return scene;
 	}
 	
-	public abstract void onUpdate(int delta);
+	public abstract void onUpdate(float delta);
 	public abstract void onDestroy();
 	public abstract void onCreate();
